Large refactors to elimit passing down redux props. Working container tests
