+++ misk/web/tabs/web-actions/src/components/index.ts
@@ -1,5 +1,5 @@
 export * from "./CommonComponents"
-export * from "./FilterWebActionsComponent"
 export * from "./RequestFormComponent"
+export * from "./RequestResponseContentTypes"
 export * from "./SendRequestCollapseComponent"
-export * from "./WebActionsComponent"
+export * from "../containers/WebActionsContainer"
