+++ misk/web/tabs/web-actions/src/containers/TabContainer.tsx
@@ -1,8 +1,13 @@
 import { H1 } from "@blueprintjs/core"
 import * as React from "react"
 import { connect } from "react-redux"
-import { WebActionsComponent } from "../components"
-import { IDispatchProps, IState, rootDispatcher, rootSelectors } from "../ducks"
+import { WebActionsContainer } from "../containers"
+import {
+  IDispatchProps,
+  IState,
+  mapDispatchToProps,
+  mapStateToProps
+} from "../ducks"
 
 class TabContainer extends React.Component<IState & IDispatchProps, IState> {
   componentDidMount() {
@@ -13,16 +18,10 @@ class TabContainer extends React.Component<IState & IDispatchProps, IState> {
     return (
       <div>
         <H1>Web Actions</H1>
-        <WebActionsComponent tag={"WebActions"} {...this.props} />
+        <WebActionsContainer tag={"WebActions"} />
       </div>
     )
   }
-}
-
-const mapStateToProps = (state: IState) => rootSelectors(state)
-
-const mapDispatchToProps = {
-  ...rootDispatcher
 }
 
 export default connect(
