+++ misk/web/tabs/web-actions/src/containers/index.ts
@@ -1,1 +1,10 @@
-export { default as TabContainer } from "./TabContainer"
+import FilterWebActionsContainer from "./FilterWebActionsContainer"
+import TabContainer from "./TabContainer"
+import WebActionCardContainer from "./WebActionCardContainer"
+import WebActionsContainer from "./WebActionsContainer"
+export {
+  FilterWebActionsContainer,
+  TabContainer,
+  WebActionCardContainer,
+  WebActionsContainer
+}
