+++ misk/web/tabs/web-actions/src/ducks/index.ts
@@ -60,6 +60,7 @@ export const rootDispatcher: IDispatchProps = {
  * State Selectors
  */
 export const rootSelectors = (state: IState) => ({
+  router: state.router,
   simpleForm: simpleRootSelector<IState, ISimpleFormImmutableState>(
     "simpleForm",
     state
@@ -96,4 +97,13 @@ export function* rootSaga(): IterableIterator<AllEffect> {
     fork(watchSimpleFormSagas),
     fork(watchSimpleNetworkSagas)
   ])
+}
+
+/**
+ * Map Dispatch/State to Props
+ */
+export const mapStateToProps = (state: IState) => rootSelectors(state)
+
+export const mapDispatchToProps: IDispatchProps = {
+  ...rootDispatcher
 }
