+++ misk/web/tabs/web-actions/src/ducks/webActions.ts
@@ -695,7 +695,11 @@ export const generateTypesMetadata = (
 
 function* handleMetadata() {
   try {
-    const { data } = yield call(axios.get, "/api/webaction/metadata")
+    // const { data } = yield call(axios.get, "/api/webaction/metadata")
+    const { data } = yield call(
+      axios.get,
+      "https://raw.githubusercontent.com/adrw/misk-web/adrw/20190409.WebActionsExampleData/examples/data/demo/webactions.json"
+    )
     const { webActionMetadata } = data
     const metadata = chain(webActionMetadata)
       .map((action: IWebActionAPI) => {
