+++ misk/web/tabs/web-actions/tests/components/commonComponents.test.tsx
@@ -1,15 +1,28 @@
 import React from "react"
 import { cleanup, render } from "react-testing-library"
-import { Metadata } from "../../src/components/CommonComponents"
+import {
+  Metadata,
+  MetadataCollapse,
+  MetadataMenu
+} from "../../src/components/CommonComponents"
+import { renderWithRedux } from "../upstreamableTestUtilities"
 
-afterEach(cleanup)
 describe("Common Components", () => {
+  afterEach(cleanup)
+  it("MetadataMenu", () => {
+    const { asFragment } = render(<MetadataMenu />)
+    expect(asFragment()).toMatchSnapshot()
+  })
   it("Metadata", () => {
     const { asFragment } = render(<Metadata content={"Test Content"} />)
     expect(asFragment()).toMatchSnapshot()
   })
-  // it("MetadataCollapse", () => {
-  //   const { asFragment } = render(<MetadataCollapse content={"Test Content"} />)
-  //   expect(asFragment()).toMatchSnapshot()
-  // })
+  it("MetadataCollapse can render with redux", () => {
+    const { asFragment } = renderWithRedux(
+      <MetadataCollapse content={"Test Content"} tag={"TestTag"}>
+        <span>{"Collapse Span"}</span>
+      </MetadataCollapse>
+    )
+    expect(asFragment()).toMatchSnapshot()
+  })
 })
